+++ install/RESOURCES/APPS/Dico/Dico_main.m
@@ -1,0 +1,30 @@
+/* 
+   Project: Dico
+
+   Author: patrick
+
+   Created: 2026-03-10 20:19:31 +0100 by patrick
+*/
+
+#import <AppKit/AppKit.h>
+
+int 
+main(int argc, const char *argv[])
+{
+// Uncomment if your application is Renaissance application
+/*  CREATE_AUTORELEASE_POOL (pool);
+  [NSApplication sharedApplication];
+  [NSApp setDelegate: [AppController new]];
+
+  #ifdef GNUSTEP
+    [NSBundle loadGSMarkupNamed: @"MainMenu-GNUstep"  owner: [NSApp delegate]];
+  #else
+    [NSBundle loadGSMarkupNamed: @"MainMenu-OSX"  owner: [NSApp delegate]];
+  #endif
+   
+  RELEASE (pool);
+*/
+
+  return NSApplicationMain (argc, argv);
+}
+